Fix PersistentVolumeClaim
Adjust accessMode to be ReadWriteOnce instead of ReadWriteMany.
By default, K3s comes with a local-path storage provision option.
This option binds persistent volumes to the node they are located on.
Therefore, it does not support ReadWriteMany, since other nodes
are not able to access the local storage of another node.