Deploy container in Helm chart with `tty: true`
This causes the container to use line buffering (as opposed to
fixed-size buffering which, if left unflushed, would often cause output
to appear delayed or not at all). Consider, for example, the following
Swift program:

    import Dispatch

    print("Hi")
    dispatchMain()

If not attached to a tty, this program will not output anything (at
least not immediately), presumably since the string does not fill the
buffer and blocking causes the program to not flush the output like it
would if the program would simply exit.

Further reading:

- https://github.com/apple/swift-nio/issues/490#issuecomment-399750811
- https://stackoverflow.com/a/42231138