Crash on disconnect
We'll let the container runtime handle reconnects for now, since it
generally supplies a better backoff strategy that `swift-discord` did
previously (reconnecting 1000 times in quick succession to the Discord
gateway might not be the best thing to do). K8s, for example, provides
an exponential backoff strategy, that might be better.